requirements.txt
requests>=2.31.0
pydub>=0.25.1
python-dotenv>=1.0.0
# ffmpeg 별도 설치 필요: brew install ffmpeg (macOS) / apt install ffmpeg (Linux)
